+++ app/server/static/components/download_seq2seq.vue
@@ -35,7 +35,7 @@ block example-format-area
 </template>
 
 <script>
-import { uploadMixin } from './mixin';
+import uploadMixin from './uploadMixin';
 
 export default uploadMixin;
 </script>
