+++ app/server/static/components/download_sequence_labeling.vue
@@ -20,7 +20,7 @@ block example-format-area
 </template>
 
 <script>
-import { uploadMixin } from './mixin';
+import uploadMixin from './uploadMixin';
 
 export default uploadMixin;
 </script>
