+++ app/server/static/components/upload_sequence_labeling.vue
@@ -40,7 +40,7 @@ block example-format-area
 </template>
 
 <script>
-import { uploadMixin } from './mixin';
+import uploadMixin from './uploadMixin';
 
 export default uploadMixin;
 </script>
